myedits

+++ CircuitsProject/Node.cpp
@@ -1,6 +1,6 @@
 #include"Node.h"
 Node::Node(int id) :ID(id) {}
 void Node::SetID(int id) { ID = id; }
-void Node::SetVoltage(complex<float> v) { voltage = v; }
+void Node::SetVoltage(complex<double> v) { voltage = v; }
 int Node::GetID() { return ID; }
-complex<float> Node::GetVoltage() { return voltage; }
+complex<double> Node::GetVoltage() { return voltage; }
